requirements.txt
tensorflow
streamlit
pillow
numpy
pytesseract
opencv-python